Improved Login and Registry and user auth through the whole app

+++ src/CartBundle/Controller/CartController.php
@@ -21,34 +21,18 @@ class CartController extends Controller
         $allGlasses = $repositoryGlasses->findAll();
 
         $repositoryCartItems = $this->getDoctrine()->getRepository(Cart::class);
-        $cartItems = $repositoryCartItems->findAll();
+        $cartItems = $repositoryCartItems->findBy(['user' => $this->getUser()]);
 
         $totalSum = 0;
 
-        foreach($cartItems as $item) {
+        foreach ($cartItems as $item) {
             $totalSum += $item->getQuantity() * $item->getProduct()->getPrice();
         }
 
         return $this->render('default/shopping-cart.html.twig', [
             'glasses' => array_slice($allGlasses, 0, 3), // gafas que te pueden interesar
             'cartItems' => $cartItems, // items del carrito,
-            'totalSum' => $totalSum 
-        ]);
-    }
-    /**
-     * @Route("/", name="delete-shopping-cart")
-     */
-    public function deleteCartItem()
-    {
-        $repositoryGlasses = $this->getDoctrine()->getRepository(Glass::class);
-        $allGlasses = $repositoryGlasses->findAll();
-
-        $repositoryCartItems = $this->getDoctrine()->getRepository(Cart::class);
-        $cartItems = $repositoryCartItems->findAll();
-
-        return $this->render('default/shopping-cart.html.twig', [
-            'glasses' => $allGlasses, // gafas que te pueden interesar
-            'cartItems' => $cartItems // items del carrito
+            'totalSum' => $totalSum
         ]);
     }
 
@@ -62,16 +46,18 @@ class CartController extends Controller
     {
         $entityManager = $this->getDoctrine()->getManager();
 
-        // Logica
-        if ($item->getQuantity() === 1) {
-            // Borrar del carrito
-            $entityManager->remove($item);
-            $entityManager->flush();
-        } else {
-            // Decrementar
-            $item->setQuantity($item->getQuantity() - 1);
-            $entityManager->persist($item);
-            $entityManager->flush();
+        if ($item->getUser() === $this->getUser()) {
+            // Logica
+            if ($item->getQuantity() === 1) {
+                // Borrar del carrito
+                $entityManager->remove($item);
+                $entityManager->flush();
+            } else {
+                // Decrementar
+                $item->setQuantity($item->getQuantity() - 1);
+                $entityManager->persist($item);
+                $entityManager->flush();
+            }
         }
 
         // Redirigir
@@ -90,11 +76,13 @@ class CartController extends Controller
     {
         $entityManager = $this->getDoctrine()->getManager();
 
-        // Logica
-        // Aumentar
-        $item->setQuantity($item->getQuantity() + 1);
-        $entityManager->persist($item);
-        $entityManager->flush();
+        if ($item->getUser() === $this->getUser()) {
+            // Logica
+            // Aumentar
+            $item->setQuantity($item->getQuantity() + 1);
+            $entityManager->persist($item);
+            $entityManager->flush();
+        }
 
         // Redirigir
         $response = $this->redirectToRoute('shopping-cart-list');
@@ -112,10 +100,12 @@ class CartController extends Controller
     {
         $entityManager = $this->getDoctrine()->getManager();
 
-        // Logica
-        // Eliminar un item en concreto
-        $entityManager->remove($item);
-        $entityManager->flush();
+        if ($item->getUser() === $this->getUser()) {
+            // Logica
+            // Eliminar un item en concreto
+            $entityManager->remove($item);
+            $entityManager->flush();
+        }
 
         // Redirigir
         $response = $this->redirectToRoute('shopping-cart-list');
